Add tests for renderDependenciesBadge
